requirements.txt
gspread==6.1.2
google-auth==2.29.0
python-dotenv==1.0.1